New idField concept on entities
Until now, identification of an entity was not needed due to internal lists. However, this kind
of internal identification based on list index is insecure, poorly concurrent and it has problem
with selected scoped operations. This is why i decided to implement an id field mark to Entities
so now, we can use both approaches, but i hardly recommend to use the new idField property in
every entity.

With this, we don't need to save the list content on each container so this also provides a huge
memory optimization.